+++ client/cody/src/completions/index.ts
@@ -12,6 +12,7 @@ import { getContext } from './context'
 import { CompletionsDocumentProvider } from './docprovider'
 import { History } from './history'
 import * as CompletionLogger from './logger'
+import { detectMultilineMode } from './multiline'
 import { CompletionProvider, InlineCompletionProvider, ManualCompletionProvider } from './provider'
 
 const LOG_MANUAL = { type: 'manual' }
@@ -112,7 +113,7 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
             return []
         }
 
-        const { prefix, suffix, prevLine: sameLinePrefix } = docContext
+        const { prefix, suffix, prevLine: sameLinePrefix, prevNonEmptyLine } = docContext
         const sameLineSuffix = suffix.slice(0, suffix.indexOf('\n'))
 
         // Avoid showing completions when we're deleting code (Cody can only insert code at the
@@ -144,7 +145,8 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
             [],
             prefix,
             suffix,
-            '\n'
+            '\n',
+            document.languageId
         )
         const emptyPromptLength = completionNoSnippets.emptyPromptLength()
 
@@ -162,11 +164,6 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
         const completers: CompletionProvider[] = []
         let timeout: number
         let multilineMode: null | 'block' = null
-        // TODO(philipp-spiess): Add a better detection for start-of-block and don't require C like
-        // languages.
-        const multilineEnabledLanguage =
-            document.languageId === 'typescript' || document.languageId === 'javascript' || document.languageId === 'go'
-
         // VS Code does not show completions if we are in the process of writing a word or if a
         // selected completion info is present (so something is selected from the completions
         // dropdown list based on the lang server) and the returned completion range does not
@@ -190,15 +187,15 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
         }
 
         if (
-            multilineEnabledLanguage &&
-            // Only trigger multiline inline suggestions for empty lines
-            sameLinePrefix.trim() === '' &&
-            sameLineSuffix.trim() === '' &&
-            // Only trigger multiline inline suggestions for the beginning of blocks
-            prefix.trim().at(prefix.trim().length - 1) === '{'
+            (multilineMode = detectMultilineMode(
+                prefix,
+                prevNonEmptyLine,
+                sameLinePrefix,
+                sameLineSuffix,
+                document.languageId
+            ))
         ) {
             timeout = 500
-            multilineMode = 'block'
             completers.push(
                 new InlineCompletionProvider(
                     this.completionsClient,
@@ -208,6 +205,7 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
                     prefix,
                     suffix,
                     '',
+                    document.languageId,
                     3,
                     multilineMode
                 )
@@ -224,6 +222,7 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
                     prefix,
                     suffix,
                     '',
+                    document.languageId,
                     2 // tries
                 )
             )
@@ -239,6 +238,7 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
                     prefix,
                     suffix,
                     '',
+                    document.languageId,
                     2 // tries
                 ),
                 // Create a completion request for the current prefix with a new line added. This
@@ -251,6 +251,7 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
                     prefix,
                     suffix,
                     '\n', // force a new line in the case we are at end of line
+                    document.languageId,
                     1 // tries
                 )
             )
@@ -322,7 +323,8 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
             [],
             prefix,
             suffix,
-            ''
+            '',
+            currentEditor.document.languageId
         )
         const emptyPromptLength = completionNoSnippets.emptyPromptLength()
 
@@ -344,7 +346,8 @@ export class CodyCompletionItemProvider implements vscode.InlineCompletionItemPr
             similarCode,
             prefix,
             suffix,
-            ''
+            '',
+            currentEditor.document.languageId
         )
 
         try {